+++ src/function/matrix/det.js
@@ -131,6 +131,6 @@ det.doc = {
         'det([-2, 2, 3; -1, 1, 3; 2, 0, -1])'
     ],
     'seealso': [
-        'diag', 'eye', 'inv', 'range', 'size', 'squeeze', 'transpose', 'zeros'
+        'concat', 'diag', 'eye', 'inv', 'range', 'size', 'squeeze', 'transpose', 'zeros'
     ]
 };